versión antes de cambiar CrudRepository por JpaRepository

+++ src/main/java/cl/portafolio/hostalapp/models/repository/IHabitacionRepository.java
@@ -3,5 +3,5 @@ package cl.portafolio.hostalapp.models.repository;
 import cl.portafolio.hostalapp.models.entity.Habitacion;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IHabitacionRepository extends CrudRepository <Habitacion, Long> {
+public interface IHabitacionRepository extends CrudRepository <Habitacion, Integer> {
 }
